+++ bitchatTests/Mocks/MockKeychain.swift
@@ -11,54 +11,57 @@ import Foundation
 
 final class MockKeychain: KeychainManagerProtocol {
     private var storage: [String: Data] = [:]
-    
+    private var serviceStorage: [String: [String: Data]] = [:]
+
     func saveIdentityKey(_ keyData: Data, forKey key: String) -> Bool {
         storage[key] = keyData
         return true
     }
-    
+
     func getIdentityKey(forKey key: String) -> Data? {
         storage[key]
     }
-    
+
     func deleteIdentityKey(forKey key: String) -> Bool {
         storage.removeValue(forKey: key)
         return true
     }
-    
+
     func deleteAllKeychainData() -> Bool {
         storage.removeAll()
+        serviceStorage.removeAll()
         return true
     }
-    
+
     func secureClear(_ data: inout Data) {
-        //
         data = Data()
     }
-    
+
     func secureClear(_ string: inout String) {
         string = ""
     }
-    
+
     func verifyIdentityKeyExists() -> Bool {
         storage["identity_noiseStaticKey"] != nil
     }
+
+    // MARK: - Generic Data Storage (consolidated from KeychainHelper)
+
+    func save(key: String, data: Data, service: String, accessible: CFString?) {
+        if serviceStorage[service] == nil {
+            serviceStorage[service] = [:]
+        }
+        serviceStorage[service]?[key] = data
+    }
+
+    func load(key: String, service: String) -> Data? {
+        serviceStorage[service]?[key]
+    }
+
+    func delete(key: String, service: String) {
+        serviceStorage[service]?.removeValue(forKey: key)
+    }
 }
 
-final class MockKeychainHelper: KeychainHelperProtocol {
-    private typealias Service = String
-    private typealias Key = String
-    private var storage: [Service: [Key: Data]] = [:]
-    
-    func save(key: String, data: Data, service: String, accessible: CFString?) {
-        storage[service]?[key] = data
-    }
-    
-    func load(key: String, service: String) -> Data? {
-        storage[service]?[key]
-    }
-    
-    func delete(key: String, service: String) {
-        storage[service]?.removeValue(forKey: key)
-    }
-}
+/// Typealias for backwards compatibility with tests using MockKeychainHelper
+typealias MockKeychainHelper = MockKeychain
